+++ LICENSE
@@ -14,4 +14,6 @@ The following files are based on the OpenPifPaf project which is
 
 - tests/test_train_mono.py
 - tests/test_train_stereo.py
+- visuals/pifpaf_show.py
+- train/losses.py
 - predict.py
