+++ monoloco/eval/geom_baseline.py
@@ -1,6 +1,5 @@
 
 import json
-import logging
 import math
 from collections import defaultdict
