+++ monoloco/train/losses.py
@@ -1,9 +1,15 @@
-"""Inspired by Openpifpaf"""
+"""
+Adapted from https://github.com/openpifpaf,
+which is: 'Copyright 2019-2021 by Sven Kreiss and contributors. All rights reserved.'
+and licensed under GNU AGPLv3
+"""
+
 
 import math
 import torch
 import torch.nn as nn
 import numpy as np
+import matplotlib.pyplot as plt
 
 from ..network import extract_labels, extract_labels_aux, extract_outputs
 
@@ -178,6 +184,58 @@ class GaussianLoss(torch.nn.Module):
             mean_values = torch.mean(values)
             return mean_values
         return torch.sum(values)
+
+
+class CustomL1Loss(torch.nn.Module):
+    """
+    Experimental, not used.
+    L1 loss with more weight to errors at a shorter distance
+    It inherits from nn.module so it supports backward
+    """
+
+    def __init__(self, dic_norm, device, beta=1):
+        super().__init__()
+
+        self.dic_norm = dic_norm
+        self.device = device
+        self.beta = beta
+
+    @staticmethod
+    def compute_weights(xx, beta=1):
+        """
+        Return the appropriate weight depending on the distance and the hyperparameter chosen
+        alpha = 1 refers to the curve of A Photogrammetric Approach for Real-time...
+        It is made for unnormalized outputs (to be more understandable)
+        From 70 meters on every value is weighted the same (0.1**beta)
+        Alpha is optional value from Focal loss. Yet to be analyzed
+        """
+        # alpha = np.maximum(1, 10 ** (beta - 1))
+        alpha = 1
+        ww = np.maximum(0.1, 1 - xx / 78)**beta
+
+        return alpha * ww
+
+    def print_loss(self):
+        xx = np.linspace(0, 80, 100)
+        y1 = self.compute_weights(xx, beta=1)
+        y2 = self.compute_weights(xx, beta=2)
+        y3 = self.compute_weights(xx, beta=3)
+        plt.plot(xx, y1)
+        plt.plot(xx, y2)
+        plt.plot(xx, y3)
+        plt.xlabel("Distance [m]")
+        plt.ylabel("Loss function Weight")
+        plt.legend(("Beta = 1", "Beta = 2", "Beta = 3"))
+        plt.show()
+
+    def forward(self, output, target):
+
+        unnormalized_output = output.cpu().detach().numpy() * self.dic_norm['std']['Y'] + self.dic_norm['mean']['Y']
+        weights_np = self.compute_weights(unnormalized_output, self.beta)
+        weights = torch.from_numpy(weights_np).float().to(self.device)  # To make weights in the same cuda device
+        losses = torch.abs(output - target) * weights
+        loss = losses.mean()  # Mean over the batch
+        return loss
 
 
 def angle_loss(orient, gt_orient):
